Hackish way to expose back a clean context
When we call `simulate` we pass the data context per reference,
and we use this context back for debugging and showing in the
explorer for example.

Until the recent "validation refactor", this context was updated
with cleaned data, but this was no more the case, because the
validation occurs while running the rules (instead of once and for
all at the start).